perf(components): pace Capacitor eager-sync below desktop
Mobile gets concurrency 1 and a smaller warm set. candidateWindow stays unbounded. Desktop and web policies are unchanged.

Model: grok-4.6
